--- packages/react-gamepad-navigation/stories/Checkerboard.stories.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { Checkbox, CheckboxProps } from '@fluentui/react-components';
-
-export const Checkerboard = (props: CheckboxProps) => <Checkbox {...props} />;
