+++ packages/react-gamepad-navigation/stories/index.stories.tsx
@@ -2,8 +2,6 @@ import { Meta } from '@storybook/react';
 import description from '../README.md';
 
 export { Default } from './Default.stories';
-export { TextButtons } from './Button.stories';
-// export { Checkerboard } from './Checkerboard.stories';
 
 const meta: Meta = {
   title: 'Packages/react-gamepad-navigation',
